Complete the implementation of auth server
