+++ components/docker-auth/cmd/authn/authentication.go
@@ -16,291 +16,67 @@
  * under the License.
  */
 
-package authn
+package main
 
 import (
-	"crypto/tls"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	"strings"
-	"time"
 
 	"github.com/cellery-io/cellery-hub/components/docker-auth/pkg/extension"
-
-	"github.com/dgrijalva/jwt-go"
 )
 
 const (
-	issuerClaim           = "iss"
-	subjectClaim          = "sub"
-	authTokenIssuerEnvVar = "REGISTRY_AUTH_TOKEN_ISSUER"
-	idpCertEnvVar         = "IDP_CERT"
-	dockerAuthCertEnvVar  = "REGISTRY_AUTH_TOKEN_ROOTCERTBUNDLE"
+	logFile = "/extension-logs/authn-ext.log"
 )
 
-func readCert(certPathEnv string, execId string) ([]byte, error) {
-	key, err := ioutil.ReadFile(os.Getenv(certPathEnv))
-	if err != nil {
-		log.Printf("[%s] Unable to read the cert : %s\n", execId, err)
-		return nil, err
-	}
-	log.Printf("[%s] Read cert successfully\n", execId)
-	return key, nil
-}
-
-func getJWTClaims(token string, execId string) jwt.MapClaims {
-	jwtToken, _ := jwt.Parse(token, nil)
-	claims, ok := jwtToken.Claims.(jwt.MapClaims)
-	if ok {
-		log.Printf("[%s] Received JWT claims successfully\n", execId)
-		return claims
-	}
-	return nil
-}
-
-func getClaimValue(claim jwt.MapClaims, claimKey string, execId string) string {
-	value, ok := claim[claimKey].(string)
-	if ok {
-		log.Printf("[%s] Received JWT claim for the claim key %s successfully\n", execId, claimKey)
-		return value
-	}
-	return ""
-}
-
-func validateToken(inToken string, cert []byte, execId string) (bool, error) {
-	publicRSA, err := jwt.ParseRSAPublicKeyFromPEM(cert)
-	if err != nil {
-		log.Printf("[%s] Error parsing the cert : %s\n", execId, err)
-		return false, err
-	}
-	token, err := jwt.Parse(inToken, func(token *jwt.Token) (interface{}, error) {
-		if _, ok := token.Method.(*jwt.SigningMethodRSA); !ok {
-			return nil, fmt.Errorf("[%s] Unexpected signing method: %s\n", execId, token.Header["alg"])
-		}
-		return publicRSA, err
-	})
-	if token != nil && token.Valid {
-		log.Printf("[%s] Token received is valid\n", execId)
-
-		return true, nil
-	}
-	log.Printf("[%s] Token received is invalid\n", execId)
-	return false, err
-}
-
-func Authenticate(uName string, token string) int {
-	log.Println("Inside authentication logic handler")
-
-	execId, err := extension.GetExecID()
-	if err != nil {
-		log.Printf("Error in generating the execId : %s\n", err)
-	}
-	log.Printf("[%s] Authentication extension reached and token will be validated\n", execId)
-	if isJWT(execId) {
-		log.Printf("[%s] Authenticate by using JWT\n", execId)
-		jwtValidity := validateJWT(token, uName, execId)
-		if jwtValidity {
-			return 0
-		} else {
-			return 1
-		}
-	} else {
-		log.Printf("[%s] Authenticate by using access token\n", execId)
-		if validateAccessToken(token, uName, execId) {
-			log.Printf("[%s] User successfully authenticated\n", execId)
-			return 0
-		} else {
-			log.Printf("[%s] User failed to authenticate\n", execId)
-			return 1
-		}
-	}
-}
-
-func validateJWT(token string, username string, execId string) bool {
-	claim := getJWTClaims(token, execId)
-	iss := getClaimValue(claim, issuerClaim, execId)
-	sub := getClaimValue(claim, subjectClaim, execId)
-
-	log.Printf("[%s] Token issuer : %s\n", execId, iss)
-	log.Printf("[%s] Subject : %s\n", execId, sub)
-
-	if sub != username {
-		log.Printf("[%s] Username(%s) does not match with subject(%s) in JWT\n", execId, username, sub)
-		return false
-	}
-
-	certificateInUse, err := readCert(idpCertEnvVar, execId)
-	if err != nil {
-		log.Printf("[%s] Unable to load idp cert file : %s\n", execId, err)
-	}
-
-	if iss == authTokenIssuerEnvVar {
-		certificateInUse, err = readCert(dockerAuthCertEnvVar, execId)
+func main() {
+	err := os.MkdirAll("/extension-logs", os.ModePerm)
+	file, err := os.OpenFile(logFile, os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	defer func() {
+		err = file.Close()
 		if err != nil {
-			log.Printf("[%s] Unable to load docker auth file : %s\n", execId, err)
+			log.Printf("Error while closing the file : %s\n", err)
+			os.Exit(2)
 		}
-	}
-
-	tokenValidity, err := validateToken(token, certificateInUse, execId)
+	}()
 	if err != nil {
-		log.Printf("[%s] Token is not valid : %s\n", execId, err)
-		return false
+		log.Println("Error creating the file :", err)
+		os.Exit(1)
 	}
-	log.Printf("[%s] Signature verified\n", execId)
+	log.SetOutput(file)
 
-	if tokenValidity {
-		log.Printf("[%s] User successfully authenticated\n", execId)
-		return true
-	} else {
-		log.Printf("[%s] Authentication failed\n", execId)
-		return false
-	}
-}
+	text := extension.ReadStdIn()
+	credentials := strings.Split(text, " ")
 
-// isJWT checks whether the token is jwt token or access token.
-func isJWT(execId string) bool {
-	isJWTEnv := os.Getenv("IS_JWT")
-	var isJWT bool
-	if len(isJWTEnv) == 0 {
-		log.Printf("[%s] Error: IS_JWT environment variable is empty\n", execId)
-		return false
-	} else {
-		if isJWTEnv == "true" {
-			log.Printf("[%s] Received a JWT token\n", execId)
-			isJWT = true
-		} else if isJWTEnv == "false" {
-			log.Printf("[%s] Received a access token\n", execId)
-			isJWT = false
-		} else {
-			log.Printf("[%s] Error: Wrong environment value given. The value should be either true or false\n",
-				execId)
-			return false
-		}
+	if len(credentials) != 2 {
+		log.Printf("Cannot parse the Input from the Auth service")
+		os.Exit(extension.ErrorExitCode)
 	}
-	return isJWT
-}
+	uName := credentials[0]
+	token := credentials[1]
 
-// validateAccessToken is used to introspect the access token
-func validateAccessToken(token string, providedUsername string, execId string) bool {
-	introspectionUrl, urlAvailability := resolveIntrospectionUrl(execId)
-	if !urlAvailability {
-		return false
-	}
-	payload := strings.NewReader("token=" + token)
-	log.Println("url", introspectionUrl)
-	req, err := http.NewRequest("POST", introspectionUrl, payload)
-	if err != nil {
-		log.Printf("[%s] Error creating new request to the introspection endpoint : %s\n", execId, err)
-		return false
-	}
-	username, password, credentialsAvailability := resolveCredentials(execId)
-	if !credentialsAvailability {
-		return false
-	}
-	req.SetBasicAuth(username, password)
-	// todo Remove the the host verification turning off
-	tr := &http.Transport{
-		TLSClientConfig: &tls.Config{
-			InsecureSkipVerify: true,
-		},
-	}
-	client := &http.Client{Transport: tr}
-	res, err := client.Do(req)
-	if err != nil {
-		log.Printf("[%s] Error sending the request to the introspection endpoint : %s\n", execId, err)
-		return false
-	}
+	log.Printf("Username : %s, Password : %s", uName, token)
+	url := fmt.Sprintf("http://localhost:8080/authentication?uName=%s&token=%s", uName, token)
+	log.Printf("Called %s", url)
+
+	req, _ := http.NewRequest("GET", url, nil)
+
+	res, _ := http.DefaultClient.Do(req)
+
 	defer res.Body.Close()
-	body, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		log.Printf("[%s] Error reading the response from introspection endpoint : %s\n", execId, err)
-		return false
-	}
-	var result map[string]interface{}
-	err = json.Unmarshal([]byte(string(body)), &result)
-	if err != nil {
-		log.Printf("[%s] Error un marshalling the json : %s\n", execId, err)
-		return false
-	}
-	isActive, ok := (result["active"]).(bool)
-	if !ok {
-		log.Printf("[%s] Error casting active to boolean. This may be due to a invalid token\n", execId)
-		return false
-	}
-	log.Printf("[%s] Resolved access token values successfully\n", execId)
-	isExpired := isExpired(result["exp"], execId)
-	isValidUser := isValidUser(result["username"], providedUsername, execId)
-	return isExpired && isActive && isValidUser
-}
 
-// resolves the IS host and port from the environment variables.
-// If the environment is not set the port and host will be resolved through the config file.
-func resolveIntrospectionUrl(execId string) (string, bool) {
-	idpEndPoint := os.Getenv("IDP_END_POINT")
-	introspectionEP := os.Getenv("INTROSPECTION_END_POINT")
-	if len(introspectionEP) == 0 {
-		log.Printf("[%s] Error: INTROSPECTION_END_POINT environment variable is empty\n", execId)
-		return "", false
-	}
-	if len(idpEndPoint) == 0 {
-		log.Printf("[%s] Error: IDP_END_POINT environment variable is empty\n", execId)
-		return "", false
-	}
-	return idpEndPoint + introspectionEP, true
-}
+	log.Printf("Response received from the auth server with the status code : %d", res.StatusCode)
 
-// resolveCredentials resolves the user credentials of the user that is used to communicate to introspection endpoint
-func resolveCredentials(execId string) (string, string, bool) {
-	username := os.Getenv("USERNAME")
-	if len(username) == 0 {
-		log.Printf("[%s] Error: USERNAME environment variable is empty\n", execId)
-		return "", "", true
+	if res.StatusCode == http.StatusUnauthorized {
+		log.Printf("Authentication failed for user %s. Exiting with error exit code", uName)
+		os.Exit(extension.ErrorExitCode)
 	}
-	password := os.Getenv("PASSWORD")
-	if len(password) == 0 {
-		log.Printf("[%s] Error: PASSWORD environment variable is empty\n", execId)
-		return "", "", true
+	if res.StatusCode == http.StatusOK {
+		log.Printf("user %s is successfully authenticated. Exiting with success exit code", uName)
+		os.Exit(extension.SuccessExitCode)
 	}
-	log.Printf("[%s] Suceesfully received credentials\n", execId)
-	return username, password, true
-}
 
-// isValidUser checks whether the provided username matches with the username in the token
-func isValidUser(tokenUsername interface{}, providedUsername string, execId string) bool {
-	if username, ok := tokenUsername.(string); ok {
-		usernameTokens := strings.Split(username, "@")
-		log.Printf("[%s] User needed to be validated %s with provided username %s\n",
-			execId, usernameTokens[0], providedUsername)
-		if providedUsername == usernameTokens[0] {
-			log.Printf("[%s] User received is valid\n", execId)
-			return true
-		}
-		log.Printf("[%s] Username does not match with the provided username %s\n", execId, providedUsername)
-	} else {
-		log.Printf("[%s] Error casting username to string. This may be due to a invalid token\n", execId)
-		return false
-	}
-	return false
-}
-
-// isExpired validated whether the username is expired
-func isExpired(timestamp interface{}, execId string) bool {
-	if validity, ok := timestamp.(float64); ok {
-		tm := time.Unix(int64(validity), 0)
-		remainder := tm.Sub(time.Now())
-		if remainder > 0 {
-			log.Printf("[%s] Token received is not expired\n", execId)
-			return true
-		}
-		log.Printf("[%s] Token received is expired. Token expiry time is %s, while the system time is %s\n",
-			execId, tm, time.Now())
-	} else {
-		log.Printf("[%s] Error casting timestamp to string. This may be due to a invalid token\n", execId)
-		return false
-	}
-	return false
 }
